+++ packages/client/src/components/noodle/NoodleView.tsx
@@ -89,15 +89,12 @@ import {
   useConfirmNoodleImagePrompts,
   useCreateNoodleInteraction,
   useCreateNoodlePost,
-  useCreatePrivateNoodleAccount,
-  useDeletePrivateNoodleAccount,
   useDeleteNoodleInteraction,
   useDeleteNoodlePost,
   useInviteNoodleCharacter,
   useInviteNoodleCharacters,
   useLoadOlderNoodlePosts,
   useNoodle,
-  useNoodlerHub,
   useRefreshNoodle,
   useRemoveNoodleCharacter,
   useRemoveNoodleInteraction,
@@ -107,15 +104,20 @@ import {
   useUpdateNoodleFillerProfile,
   useRescheduleNoodleRefresh,
   useResetNoodleTimeline,
-  useRetryPrivateIdentityGeneration,
-  useSubscribeNoodleAccount,
-  useUnlockNoodlePost,
-  useUnsubscribeNoodleAccount,
   useUpdateNoodleAccount,
   useUpdateNoodleInteraction,
   useUpdateNoodlePost,
   useUpdateNoodleSettings,
 } from "../../hooks/use-noodle";
+import {
+  useCreatePrivateNoodleAccount,
+  useDeletePrivateNoodleAccount,
+  useNoodlerHub,
+  useRetryPrivateIdentityGeneration,
+  useSubscribeNoodleAccount,
+  useUnlockNoodlePost,
+  useUnsubscribeNoodleAccount,
+} from "../../hooks/use-noodler";
 import { useUIStore } from "../../stores/ui.store";
 
 type RawCharacter = { id?: unknown; data?: unknown; avatarPath?: unknown };
